pyproject.toml
[tool.poetry]
name = "rag-trials-public"
version = "0.1.0"
description = ""
authors = ["rishabh <rishabh.tomar@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
llama-index-llms-openai = "^0.1.22"
llama-index-experimental = "^0.1.3"
streamlit = "^1.35.0"
openpyxl = "^3.1.4"
streamlit-tailwind = "^0.1.0"


[tool.poetry.group.dev.dependencies]
setuptools = "^70.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
